Add example json payloads as sent by Chromium
I recorded this by going through the register and login flow, and
dumping the request bodies. We can use this data to write tests for
FromJSON instances, they should be able to handle this.

We'll put them in a directory named after the endpoint, so the test can
check all files in the directory.