Drop js dependencies and build tools
This adds a few lines of JavaScript to handle base64url encoding and
decoding of the array buffers that the browser returns/expects, so we
can convert to the json format that the server expects.

Previously this used a third-party library, but that requires bringing in
Yarn and Packer to manage this, which results in a transitive dependency
on almost 600k lines of JavaScript (excluding blank lines and comments)
and another 9k lines of TypeScript.

By implementing the conversion by hand, we reduce the footprint of the
example by:

 * Dropping the 64 MB node_modules directory.
 * Reducing the size of the dev dependencies Nix closure by 100 MB.
 * Removing the 6.4 seconds required to run Yarn initially.
 * Removing the 4.3 seconds required to run Parcel initially.